==> notebooks/2_merge_data.ipynb <==
### Preprocesing of accident and incident data
- merge datasets
- add weather data
- filter data

import pandas as pd
import numpy as np
import datetime
import csv
import seaborn as sns
import matplotlib.pyplot as plt

paths = [
    "../data/1_raw/txt/accident_gravity_reporting_regulated_jul2020_present.txt",
    "../data/1_raw/txt/accident_hazardous_liquid_jan2010_present.txt",
    "../data/1_raw/txt/incident_gas_distribution_jan2010_present.txt",
    "../data/1_raw/txt/incident_gas_transmission_gathering_jan2010_present.txt",
    "../data/1_raw/txt/incident_liquefied_natural_gas_jan2011_present.txt",
    "../data/1_raw/txt/incident_type_r_reporting_regulated_gas_gathering_may2022_present.txt"
]

data_sources = [
    'gravity',
    'hazardous_liquid',
    'gas_distribution',
    'gas_transmission',
    'liquefied_natural_gas',
    'gas_gathering'
]

# concatenate tables to one df

df_temp1 = pd.read_csv(
    paths[0]
    , header=0
    , encoding='unicode_escape' 
    , sep='\t'
    #, index_col=['REPORT_NUMBER', 'SUPPLEMENTAL_NUMBER'] 
    , engine='python'
    )
df_temp1['data_source'] = data_sources[0]

df_temp2 = pd.read_csv(
    paths[1]
    , header=0
    , encoding='unicode_escape' 
    , sep='\t'
    #, index_col=['REPORT_NUMBER', 'SUPPLEMENTAL_NUMBER'] 
    , engine='python'
    )
df_temp2['data_source'] = data_sources[1]

df = pd.concat([df_temp1, df_temp2])

for path, data_source in zip(paths[2:], data_sources[2:]):
    df_temp = pd.read_csv(
        path
        , header=0
        , encoding='unicode_escape' 
        , sep='\t'
        #, index_col=['REPORT_NUMBER', 'SUPPLEMENTAL_NUMBER'] 
        , engine='python'
        )
    df_temp['data_source'] = data_source
    # convert mft to bbls
    df_temp['INTENTIONAL_RELEASE_BBLS'] =df_temp['INTENTIONAL_RELEASE'] * 178.10760668
    df_temp['UNINTENTIONAL_RELEASE_BBLS'] =df_temp['UNINTENTIONAL_RELEASE'] * 178.10760668  
    # drop columns
    df_temp.drop(labels=['INTENTIONAL_RELEASE', 'UNINTENTIONAL_RELEASE'], axis=1, inplace=True)
    df = pd.concat([df, df_temp])

print(df.shape)

# create columns 'case_date', 'YYYY-mm', 'YYYY-mm_prev'
df2 = df.copy()
df2['LOCAL_DATETIME'] = pd.to_datetime(df2['LOCAL_DATETIME'])
df2['case_date'] = df2['LOCAL_DATETIME'].dt.date
df2['YYYY-mm'] = pd.to_datetime(df2['case_date']).dt.strftime('%Y-%m')
df2['YYYY-mm_prev'] = (pd.to_datetime(df2['YYYY-mm']) - pd.DateOffset(months=1)).dt.strftime('%Y-%m')

df = df2.copy()

# drop unnamed columns
columns_to_drop = df.columns[df.columns.str.contains('Unnamed')]
df = df.loc[:, ~df.columns.isin(columns_to_drop)]

# rename some columns
df.rename(columns={'LOCATION_LATITUDE': 'case_lat', 'LOCATION_LONGITUDE': 'case_lon'}, inplace=True)

# drop dates outside of our range
# and records with no locations
df = df[
    (df['case_date'] >= datetime.date(2010, 1, 1)) 
    & (df['case_date'] <= datetime.date(2022, 12, 31))
    & (df['case_lat'].notnull())
    & (df['case_lon'].notnull())
    ]

# drop extreme locations
df = df[(df['case_lon'] <= -70) & (df['case_lon'] >= -152)]

# reset_index
df.reset_index(inplace=True, drop=True)

"""
# export to qgis for further processing
cols_qgis = [
             #'case_idx',
             'case_date',
             'case_lat',
             'case_lon',
             'COMMODITY_RELEASED_TYPE',
]

df_qgis = df[cols_qgis].copy()
# rename some columns
df_qgis.rename(columns={'COMMODITY_RELEASED_TYPE': 'rlsd_type'}, inplace=True)

df_qgis.to_csv('../data/1_raw/meteo/df_qgis.csv', index=True, index_label='case_idx')
#df_qgis.to_csv('../data/1_raw/meteo/df_qgis.csv', index=False)

"""
"""
PROCESSING IN QGIS:
    - assign nearest station to each observation
    - assign climate zone to each observation
"""

# import data on zones and weather
df_qgis_stations_zones = pd.read_csv('../data/1_raw/meteo/df_qgis_stations_zones.csv', header=0)

df_qgis_stations_zones.drop(['zones1_BA_Climate', 'zones2_BA_Climate', 'zones3_BA_Climate'], axis=1, inplace=True)

df_weather = pd.read_csv('../data/1_raw/meteo/weather.csv', header=0)

df_qgis_stations_zones.head()

df_weather.head()

# add station and zone data to df
df = pd.merge(df, df_qgis_stations_zones[['case_idx', 'station_id', 'zone']], how='left', left_index=True, right_on='case_idx')

df.head()

# add weather data to df
df_ = pd.merge(df, df_weather[['YYYY-mm', 'station_id', 'TAVG']], how='left', on=['YYYY-mm', 'station_id'])
df2 = pd.merge(df_, df_weather[['YYYY-mm', 'station_id', 'TAVG']], how='left', left_on=['YYYY-mm_prev','station_id'], right_on=['YYYY-mm', 'station_id'])

df2.drop_duplicates(inplace=True)

df2.drop('YYYY-mm_y', axis=1, inplace=True)
df2.rename(columns={'Y-mm_x': 'YYYY-mm'}, inplace=True)

df2.head()

mapping = {'YES': 1, 'NO': 0}

cols_binary = [
    'IGNITE_IND',
    'EXPLODE_IND',
    'FEDERAL',
    'CROSSING',
    'COULD_BE_HCA',
    'SCADA_IN_PLACE_IND',
    'EMPLOYEE_DRUG_TEST_IND',
    'CONTRACTOR_DRUG_TEST_IND',]

df2[cols_binary] = df2[cols_binary].map(lambda x: mapping.get(x,x))

df2['case_date'] = pd.to_datetime(df2['case_date'])

# calculate installation age in days
df2['INSTALLATION_YEAR'].replace('UNKNOWN', np.nan, inplace=True)
df2['INSTALLATION_YEAR'] = pd.to_datetime(df2['INSTALLATION_YEAR'])

df2['inst_age_in_days'] = (df2['case_date'] - df2['INSTALLATION_YEAR']).dt.days

df2.loc[:, ['TAVG_x', 'TAVG_y']] = df2.loc[:, ['TAVG_x', 'TAVG_y']].fillna(0)
df2['TAVG'] = (df2['TAVG_x'] + df2['TAVG_y']) / 2
df2.drop(labels=['TAVG_x','TAVG_y'], axis=1, inplace=True)

# separate procedure to fill nans in TAVG using QGIS - by finding next nearest station

# import df of filled nans
df_filled_nans = pd.read_csv('../data/1_raw/meteo/df_filled_nans.csv', header=0)

# calculate avg temp from TAVG and TAVG_y

df_filled_nans.loc[:, ['TAVG', 'TAVG_y']] = df_filled_nans.loc[:, ['TAVG', 'TAVG_y']].fillna(0)

df_filled_nans['TAVG_x'] = (df_filled_nans['TAVG'] + df_filled_nans['TAVG_y']) / 2

# fill nans if df2
df2['TAVG'] = df2['TAVG'].fillna(df_filled_nans.set_index('case_idx')['TAVG_x'])

df2['TAVG'].isna().sum()

# check nans in ACCIDENT_PSIG and MOP_PSIG
print(df2['ACCIDENT_PSIG'].isna().sum()/df2['ACCIDENT_PSIG'].shape[0])
print(df2['MOP_PSIG'].isna().sum()/df2['MOP_PSIG'].shape[0])

df2['accident_pressure_as_%_mop_psig'] = df2['ACCIDENT_PSIG'] / df2['MOP_PSIG'] * 100
print(df2['accident_pressure_as_%_mop_psig'].isna().sum()/df2['accident_pressure_as_%_mop_psig'].shape[0])

filtr = (df2['ACCIDENT_PSIG'] == 0) & (df2['MOP_PSIG'] == 0)
df2.loc[filtr, 'accident_pressure_as_%_mop_psig'] = 100

print(df2['accident_pressure_as_%_mop_psig'].isna().sum()/df2['accident_pressure_as_%_mop_psig'].shape[0])

filtr = (df2['ACCIDENT_PSIG'].isna()) | (df2['MOP_PSIG'].isna())
df2.loc[filtr, 'accident_pressure_as_%_mop_psig'] = 100

print(df2['accident_pressure_as_%_mop_psig'].isna().sum()/df2['accident_pressure_as_%_mop_psig'].shape[0])

filtr = (df2['ACCIDENT_PSIG'] > 0) & (df2['MOP_PSIG'] == 0)
df2.loc[filtr, 'accident_pressure_as_%_mop_psig'] = 100

df2[['ACCIDENT_PSIG', 'MOP_PSIG', 'accident_pressure_as_%_mop_psig']].to_csv('processed_data_test.csv', index=True, index_label='case_idx')

nan_share_per_column = df2.isna().sum()/df2.shape[0]

cols_20prcnt_nan = nan_share_per_column[nan_share_per_column <= 0.2].index.tolist()
df_20prcnt_nan = df2.loc[:, cols_20prcnt_nan]

# manually selected columns
selected_cols = [
    'data_source',    
    'case_lat',
    'case_lon',
    'case_date',
    'INSTALLATION_YEAR',
    'COMMODITY_RELEASED_TYPE',
    'INTENTIONAL_RELEASE_BBLS',
    'UNINTENTIONAL_RELEASE_BBLS',    
    'FATAL',
    'INJURE',
    'ON_OFF_SHORE',
    'IGNITE_IND',
    'EXPLODE_IND',
    'NUM_PUB_EVACUATED',
    'FEDERAL',
    'LOCATION_TYPE',
    'CROSSING',
    'ITEM_INVOLVED',
    'MATERIAL_INVOLVED',
    'EST_COST_OPER_PAID',
    'EST_COST_GAS_RELEASED',
    'EST_COST_PROP_DAMAGE',
    'EST_COST_EMERGENCY',
    'EST_COST_ENVIRONMENTAL',
    'EST_COST_OTHER',
    'CAUSE',
    'CAUSE_DETAILS',
    'NARRATIVE',
    'SYSTEM_PART_INVOLVED',
    'INCIDENT_AREA_TYPE',
    'PIPE_FACILITY_TYPE',
    'inst_age_in_days',
    'RELEASE_TYPE',
    'COULD_BE_HCA',
    'ACCIDENT_PSIG',
    'MOP_PSIG',
    'accident_pressure_as_%_mop_psig',
    'PIPELINE_FUNCTION',
    'SCADA_IN_PLACE_IND',
    'INVESTIGATION_STATUS',
    'EMPLOYEE_DRUG_TEST_IND',
    'CONTRACTOR_DRUG_TEST_IND',
    'zone',
    'TAVG',
]

df2 = df2.loc[:, selected_cols]

print(df2.isna().sum()/df2.shape[0])

df2.to_csv('../data/2_merged/merged_data.csv', mode='w', index=True, index_label='case_idx')
