==> notebooks/4_one_hot.ipynb <==
import time
import datetime

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

SEED = 260420010

df = pd.read_csv('../data/3_fill_nans/processed_data.csv')
df = df.sample(frac=1, replace=False, random_state=SEED)

# One Hot Encode

one_hot_col = ["data_source", "COMMODITY_RELEASED_TYPE", "ON_OFF_SHORE", "LOCATION_TYPE", 
               "ITEM_INVOLVED", "MATERIAL_INVOLVED", "SYSTEM_PART_INVOLVED", 
               "INCIDENT_AREA_TYPE", "PIPE_FACILITY_TYPE", "PIPELINE_FUNCTION", "INVESTIGATION_STATUS",
               "zone", 'RELEASE_TYPE', 'CAUSE_DETAILS', 'CAUSE']
# one_hot_col = ["COMMODITY_RELEASED_TYPE", "ON_OFF_SHORE", "LOCATION_TYPE", 
#                "MATERIAL_INVOLVED", 
#                "PIPE_FACILITY_TYPE", "PIPELINE_FUNCTION",
#                "zone"]

test = 0

for col in one_hot_col:
    unique_features = df[col].nunique()
    test += unique_features
    print(f"Ilość unikalnych cech w kolumnie '{col}': {unique_features}")
print(test - len(one_hot_col))

def one_hot_encode(input_dataframe: pd.DataFrame):
    for col in one_hot_col:
        one_hot_encoded = pd.get_dummies(input_dataframe[col], dummy_na=True, prefix=col, dtype=int)
        null_cols = []
        for oh_enc_name, oh_enc_col in one_hot_encoded.items():
            if oh_enc_col.sum() == 0:
                null_cols.append(oh_enc_name)
        input_dataframe = pd.concat([input_dataframe, one_hot_encoded], axis=1)
        input_dataframe.drop(null_cols + [col], axis=1, inplace=True)
    return input_dataframe


df_one_hot = one_hot_encode(df)

nan_columns = df_one_hot.columns[df_one_hot.isna().any()].tolist()

if len(nan_columns) == 0:
        print("Nie ma wartości NaN w żadnej kolumnie.")
else:
    print("Kolumny zawierające wartości NaN:")
    print(nan_columns)

    # Obliczanie procentowego udziału wartości NaN w każdej kolumnie
    for col in nan_columns:
        nan_percentage = (df_one_hot[col].isna().sum() / len(df)) * 100
        print(f"Procent NaN w kolumnie {col}: {nan_percentage:.2f}%")

df_one_hot.to_csv(f'../data/4_one_hot/processed_data.csv', index=False)
