==> notebooks/5_tf_idf.ipynb <==
W tym notebooku chcemy mieć kod, który robi tf_idf z Narrative, łączy wynik z pozostałymi danymi i generuje gotowe dane do modelu.

import numpy as np
import pandas as pd
import nltk
import spacy
from matplotlib import pyplot as plt
from tqdm import tqdm

from sklearn.feature_extraction.text import TfidfVectorizer,CountVectorizer
from sklearn.decomposition import PCA
from statsmodels.graphics.gofplots import qqplot
from scipy.stats import norm, uniform
from pywsd.utils import lemmatize_sentence

pd.set_option('display.max_columns', None)
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')


df = pd.read_csv('../data/4_one_hot/processed_data.csv')

text_data = df.loc[:, df.dtypes == 'object']

df['NARRATIVE'].isna().sum()

df['NARRATIVE'] = df['NARRATIVE'].fillna('')

nlp = spacy.load('en_core_web_sm')
narrative_sentences = []
for raw_sent in tqdm(df['NARRATIVE']):
    doc = nlp(''.join([word.lower() for word in raw_sent]))
    lemmatized_tokens = [token.lemma_ for token in doc]
    lemmatized_text = ' '.join(lemmatized_tokens)
    narrative_sentences.append(lemmatized_text)

tfidfvectorizer = TfidfVectorizer(analyzer='word',stop_words= 'english')
tfidf_wm = tfidfvectorizer.fit_transform(narrative_sentences)
tfidf_tokens = tfidfvectorizer.get_feature_names_out()
df_tfidfvect = pd.DataFrame(data = tfidf_wm.toarray(),columns = tfidf_tokens)

df_tfidfvect.sum(axis=0).sort_values(ascending=False).head(20)

pca = PCA(svd_solver='randomized', n_components=500)
reduced_tfidf = pca.fit_transform(df_tfidfvect)

reduced_tfidf.shape

column_names = [f'narrative_tfidf-PC-{i}' for i in range(500)]
tfidf = pd.DataFrame(reduced_tfidf, columns=column_names)
tfidf.to_csv('../data/5_tf_idf/tf-idf.csv', index=False)
