==> notebooks/6_fill_with_predictions.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.tree import DecisionTreeRegressor
from sklearn.linear_model import LinearRegression
from sklearn.metrics import r2_score, mean_squared_error
SEED = 260420010

df = pd.read_csv('../data/4_one_hot/processed_data.csv').drop(['NARRATIVE'], axis=1)

def fill_with_predictions(model, df: pd.DataFrame, columns):
    df = df.copy()
    for col in columns:
        X_train, y_train = df.drop([col], axis=1).loc[np.logical_not(df[col].isna())], df[col].loc[np.logical_not(df[col].isna())]
        X_test = df.drop([col], axis=1).loc[df[col].isna()]
        model.fit(X_train, y_train)
        df.loc[df[col].isna(), col] = model.predict(X_test)
    return df

columns_to_fill_with_predict = ['inst_age_in_days']
filling_model = LinearRegression()
df = fill_with_predictions(filling_model, df, columns_to_fill_with_predict)

df.to_csv('../data/6_fill_with_predictions/processed_data.csv', index=False)
