==> notebooks/7_drop_columns.ipynb <==
## DROP COLUMNS

We need to drop everything odd, except target variable here, to set up `target` to the predictable column and `with_narrative_tf_idf` to `True`, if you need a `NARRATIVE` tf-idf matrix.

import pandas as pd

target = 'CAUSE'
with_narrative_tf_idf = False

def drop_columns(df: pd.DataFrame, drop_columns):
    df = df.copy()
    for drop_col in drop_columns:
        to_drop_iter = []
        for col_name, _ in df.items():
            if col_name.startswith(drop_col):
                to_drop_iter.append(col_name)
        df = df.drop(columns=to_drop_iter)
    return df

columns_to_drop=['INTENTIONAL_RELEASE_BBLS', 'UNINTENTIONAL_RELEASE_BBLS', 'FATAL','INJURE', 'IGNITE_IND','EXPLODE_IND','NUM_PUB_EVACUATED', 'ITEM_INVOLVED',
                                                    'EST_COST_OPER_PAID', 'EST_COST_GAS_RELEASED', 'EST_COST_PROP_DAMAGE', 'EST_COST_EMERGENCY', 'EST_COST_ENVIRONMENTAL', 'EST_COST_OTHER',
                                                    'CAUSE_DETAILS',	'CAUSE', 'NARRATIVE',	'SYSTEM_PART_INVOLVED',	'INCIDENT_AREA_TYPE',
                                                    'RELEASE_TYPE', 'ACCIDENT_PSIG', 'accident_pressure_as_%_mop_psig', 'SCADA_IN_PLACE_IND', 'INVESTIGATION_STATUS', 'EMPLOYEE_DRUG_TEST_IND',
                                                    'CONTRACTOR_DRUG_TEST_IND', 'case_idx', 'data_source', 'INSTALLATION_YEAR', 'CROSSING', 'COULD_BE_HCA']

columns_to_drop.remove(target)
df = pd.read_csv('../data/6_fill_with_predictions/processed_data.csv')
if with_narrative_tf_idf:
    tf_idf = pd.read_csv('../data/5_tf_idf/tf-idf.csv')
    df = pd.concat([df, tf_idf], axis=1)
df_clean = drop_columns(df, columns_to_drop)

def reverse_one_hot_encoding(df: pd.DataFrame, col_to_revert):
    df_copy = df.copy()
    for col_name in df.columns:
        if col_name.startswith(col_to_revert):
            col_value = col_name.replace(col_to_revert+'_', '')
            df_copy.loc[df[col_name] == 1, col_to_revert] = col_value
            df_copy = df_copy.drop([col_name], axis=1)
    return df_copy

df_clean = reverse_one_hot_encoding(df_clean, target)
df_clean.to_csv('../data/7_dropped_unnecessary_columns/processed_data.csv', index=False)
